Add minpoly() to compute minimal polynomial
Only works over base field or over subfield of which it is a member. Requires matrix.py, which is imported.
Added __int__ to return field element's polynomial representation as an integer, which for ffield(p) is the mod p number itself. 
Other minor changes, such as using int.bit_length when available.