Major overhaul of transcendental functions
exp, ln, sin, atan are now computed by Taylor series after range reduction. Series are terminated based on SIGNIFICANCE, which is set to 80 (bits). But constants have at least 256 bits of significance.
Added approximate(accuracy) instance method to limit the magnitude of denominators (or numerators for numbers in (-1,1)). This method is applied with accuracy=ACCURACY to the arguments and results of the series computations. ACCURACY=1<<256.
Removed log2 instance method; it is now subsumed by the log instance method, which has always taken a base as an optional arg. Added exp instance method.
__pow__ no longer needs tricks to increase bits of significance, as log and exp have sufficient accuracy.