Significant changes to handle nans and infs
nans and infs can't be converted to rational, so in these cases operations have to try to instead convert rational or xrational to float or complex.
Meanwhile, added a lot of the functions that are found in the math module, but using rational and xrational instead of float or complex. exp, log, trig and hyperbolic trig functions have been added, and are defined for the entire complex plane.