Allowed a/b string input format, e.g. '+2/3'
Numerator and denominator can both be in bstr format, but sign and base# can only appear in numerator, and denominator's base is assumed to be that of numerator. For complex rationals, real and imaginary parts are independent. Example: '3/5-4#1/11i'
Note that it is recommended to use '*i' rather than just 'i' when the base is >10, and required for base > 18, although substituting white space for * is accepted.