Made erfc accurate; improved performance
Added an arg, _gcd_, to rational() intended for internal use and defaulting to 1. If 0, the gcd(a,b) step in creating a rational is skipped, so gcd(a,b) is implicitly guaranteed to be 1. Certain instance methods use this feature to avoid an unnecessary gcd step: approximate, abs, neg, pow with integer power or exact roots.
Changed range to xrange wherever the former appeared. (Irrelevant for python3)
Added an internal _erf routine which takes a second argument indicating the extra precision required [for computing erfc as 1-erf]. erf calls _erf except for special args such as nan and inf; erfc calls _erf if its arg is small positive or the erfc series starts diverging before providing sufficient precision.