Fixes due to qrational cases
Since the rational, xrational, and qrational constructors may return a different class in some cases, binary operators need to be careful when trying to convert other to self's class and not assume the conversion actually happened. In particular, rational() may return an xrational or qrational, and xrational() may return a qrational. To see if an xrational or qrational (or anything else) is successfully converted to a rational: x = rational(x); if x is x.real ....