Support for Gaussian, Lipschitz, Hurwitz integers
In the rational, xrational, and qrational classes, added methods gaussian, lipschitz, and hurwitz, to respectively round to the nearest "integer" of the corresponding kind. For rational and xrational, the three methods are identical.
For xrational and qrational classes, floordiv, mod, divmod methods and their variants have been added. The result of a//b is (a/b).hurwitz(). This differs from the behavior of the corresponding methods for python2 complex types. There are no corresponding methods for python3. 
xgcd is a new method that performs the extended gcd algorithm. xgcd(a,b) returns (g,c,d), where g is a greatest common divisor of a and b, and is equal to ca+db. It uses the Euclidean algorithm over the appropriate "integers", i.e., normal integers for the rational class, Gaussian integers for the xrational class, and Hurwitz integers for the qrational class. Note that qrational multiplication is not commutative.