xgcd now returns positive integer gcd if possible
a and b are relatively prime exactly when xgcd(a,b)[0]==1. But a gcd remains a gcd when multiplied by a unit.
For integers, the only units are 1 and -1. For Gaussian integers, there are 2 additional units: i and -i.
For Hurwitz integers, there are 20 additional units: j  and -j, k and -k, and the 16 quaternions with 1/2 or -1/2 in each component.
Note that quaternion multiplication is not commutative. But a*b = ~(~b*~a) or equivalently ~(a*b) = ~b*~a, where ~ is conjugate.
We've implemented xgcd where if g,c,d=xgcd(a,b), then g = c*a+d*b, and so a/g and b/g are Hurwitz integers when a and b are.
To get g,c,d such that g = a*c+b*d and where (1/g)*a and (1/g)*b are Hurwitz integers, we note that ~g = ~c*~a + ~d*~b, so
~g,~c,~d = xgcd(~a,~b). In other words, apply xgcd to the conjugates and then take the conjugates of the results. Of course, this is unnecessary when a and b are Gaussian integers, as complex multiplication is commutative. And if you only care about relative primeness, none of this matters, since if gcd(a,b)==1, gcd(~a,~b)==1 as well. (For rational, xrational, and qrational, ~ is the conjugate operator; not compatible with int's ~, which is the one's complement operator. Note that xgcd works fine with int operands, and the result is then a triple of ints.)