Created stradix method; avoided limit on len()
ffield __len__ can legitimately return a value too large for len(), which seems to have an enforced implementation-dependent limit for some poorly-justified reason. So rather than len(F), we're doing F.__len__(), which hopefully will continue to avoid the issue.
Also, we've avoided the python2 limit on xrange by replacing it with a method that returns a generator if the built-in xrange raises an OverflowError exception.
We've broken out from ffield __str__ a method called stradix, which returns a string that is the specified radix representation of the specified int, with a specified minimum length produced by padding with 0s on the left.
The hexify method in share.py now uses stradix, and can handle bases other than 16.