Added methods for demo
For compatibility with secretsharing.html presentation, uncommented default definition of F and restored hexify(F).
Added selectshares method to select specified subset of shares.
Added mungeshare method to modify a single share in a set of k shares to produce a specified secret for that set. This demonstrates that the k-1 unmodified shares give no information about the secret.