Updated signature of ffield instances
If F is an ffield instance, the following are how F(x) is interpreted to create an element of the field:
x is callable: x(p) is evaluated and the result replaces x, which must be one of the following cases:
x is an integer: Must have -p**n < x < p**n. If x < 0, result is -F(x). Otherwise the result is the element with _x attribute x. Note: this is different behavior for negative x; the old behavior is still available as F[x].
x is a string: p must not exceed 36. The string is interpreted as a number in base p, using zits from the first p characters in 012...89a...z. The resulting number must be < p**n and becomes the _x attribute.
x is an iterable: In this case the resulting sequence must be a sequence of integers with absolute values < p. The integers, taken mod p, are considered the coefficients of a polynomial, with constant coefficient last. The polynomial is evaluated at p, and the result, which must be < p**n, becomes the _x attribute.

Tests for these cases have been added to test_ffield.py