Allow rational integers; more liberal equality
For operators, anywhere an integer is allowed now also allows a rational equal to an integer.
The method rint, when given a rational equal to an integer, returns that integer, otherwise it is the identity function.
self == other is True when the types are identical and the values are identical, but also if 0 <= self._x < self._p and self._x == other. So elements in the base GF(p) subfield compare equal with corresponding integers; in particular, 0 and 1. This also means that elements in finite fields with different characteristics will compare equal if they're in their respective base subfields and have a common value. The ffield hash method has been modified accordingly, so equal objects hash equal. Note that just because a==b and c==d doesn't mean that a+c == b+d or a*c == b*d or ...., and one side may even raise an exception. But that equality wasn't ever always true in Python anyway, even for normal numeric types (like ints and floats).