Adjust arg before approximating digamma
For computation of digamma in the general case, use same technique as for lgamma, incrementing the real part of arg until it's at least 32. Previous version used maxnorm rather than real part, but that is a slower way to do the same thing unless the absolute value of the imaginary part of the arg is at least 32. The right thing to do is probably to use abs2 and compare with 1024, but that would be even slower.