option: least irreducible Z_2 poly with fewest 1s
In class ffield creation, specifying fields of characteristic 2 when not poly and poly != None now uses the field representation whose modulus is the lexicographically least irreducible Z_2 polynomial with the fewest nonzero terms. Certain cryptographic standards specify this representation.
If poly is None or not specified, the field representation whose modulus is the lexicographically least irreducible polynomial is used. In previous versions, this was also the result whenever not poly.