Added __delitem__ to matrix and bmatrix classes
Can only delete a slab or set of slabs, i.e. a subset where all but one of the indices is full range. So, for a 2D matrix, a row or a set of rows or a column or a set of columns.
Updated ffield minpoly method to use this new functionality.