isprimitive now consistent with isirreducible
isprimitive, like isirreducible, now expects leading term to be elided and assumes it to be 1.
Also, isprimitive no longer builds huge polynomial dividends. Instead, it uses a dividend of all 1s but adds the 1s on the fly as it divides, continually checking for a remainder of 0. The eliminates the need to do more than one divide (one for each prime factor of p^n-1), but requires an initial check for x-1.