Mild optimization to matrix inverse calculation
By concatenating an identity matrix to the matrix being inverted rather than having the identity matrix first, we don't need to modify elements in or left of the current column as we normalize a row (by dividing by the pivot element) and zero the current column below the diagonal (by subtracting a multiple of the pivot row). (We just presume the diagonal element is now 1 with 0s below it.) This saves a little bit of work.