Added rffield.py for finite field extensions
Class ffieldx (in rffield.py) creates a finite field extension class given a monic irreducible polynomial over a finite field to be used as a modulus. This produces a finite field class whose instances are represented as packed polynomials over the subfield containing the coefficients of the modulus. The coefficients of each element's polynomial representation use the representation of the subfield, which may be an ffield or an ffieldx. (Hence r for recursive.)
test_ffield.py has an option -r to test ffieldx instances.