Added id property for finite field types
Removed _ffieldx dictionary; _ffield dictionary now includes both ffield and ffieldx types. id property is key for that dictionary, so if F is a finite field type _ffield[F.id] is F.
There was actually a bug in the indexing system for _ffieldx, as the key did not contain sufficient info to uniquely identify a ffieldx type.
Added _poly attribute to ffieldx, giving the packed, elided polynomial modulus with each coefficient represented as in the subfield (._basefield).
Updated the hash functions for finite field types.