Updated arithmetic algorithms and tests
Avoided using lists to unpack, compute, and repack polynomials; instead, unpack, compute, and pack are done coefficient by coefficient. Instead of using Horner's method for repacking, which requires using the list produced by unpacking in reverse order, we keep track of the current power of p as we unpack. We also use divmod rather than separately doing // and %, even though we still have to do another %, but more of the arithmetic is done with mod p numbers. And rather than unconditionally doing n iterations, we terminate as soon as the arg(s) shrink to 0. (We could further optimize as soon as one of the args shrinks to 0, but we haven't done that.)
Tests now include combinations of compatible field instances, e.g., a+b where a is an instance of ffield F, and b an instance of F.subfield or even F.subfield.subfield