Use realize to revert rationalize
nzpolydivrem and nzpolymod rationalize their arguments' coefficients if all are int, float, complex, rational or xrational, and any are int, float or complex,
and then realize the coefficients of the result(s). rationalize means applying rational; realize means applying realize.
While this requires checking all the coefficients, it means that the results of divmod, //, and % more closely resemble the results of / (which always rationalizes its args).