Added ffactor & ffactors to memoize factorizations
Efficient computation of element order in GF(q) requires factorization of q-1. For large fields, this can be time-consuming.
Similarly, efficiently checking of irreducible polynomials over GF(q) to see if they are primitive requires factorization of q^d-1, where d is the degree of the polynomial.
So, we've added ffactor(n), which is like factor(n), but assembles all the factors into a tuple, saves the tuple, and returns it. Both factor(n) and ffactor(n) first check to see if the factors of n have already been memoized, and if so, use the memoized results. Similarly ffactors(n) calls ffactor(n), whereas factors(n) calls factor(n).
We've replaced factor(n) with ffactor(n) for element order computation in ffield.py and for isprimitive and isprimitive2 methods in numfuns.py and poly.py
We've added a new signature for class ffield--if poly is specified as an empty tuple, i.e. F = ffield(p**n,()) or ffield(p,n,()), with n > 1, the lexicographically least primitive polynomial is used as the modulus. This means that F.generator.x == p. Note that if n = 1, all degree-1 polynomials are irreducible and produce the same representation, so we always choose the polynomial x as the modulus, independent of any correctly specified degree-1 polynomial modulus.