Fix factor() finding linear factors
Factors of a polynomial whose coefficients are all of type rational must have type rational coefficients. For a polynomial whose coefficients are a mix of type int, float, and rational, the coefficients are all mapped to rational type before factoring, and then the factors are mapped to polynomials with int and if necessary float coefficients.