Replace Q[x] factoring algorithm
New factoring algorithm factors polynomial over Z[x] by mapping polynomial to GF(p)[x] for suitable primes p and mapping the resulting factors back to Z[x] when they work.
This is a probabilistic algorithm in the sense that it only tries a limited number of primes before giving up in order to deal with irreducible polynomials such as x^4+1 which factor over every GF(p). So it's possible that one of the factors it finds might not be irreducible.