Minor optimizations
Changed polynomial's __hash__ to avoid calling its __getitem__
Updated Q[x] factoring to avoid retrying a given combination of GF(p)[x] factors (which otherwise could happen when a combination turns into a Q[x] factor, the combination's GF(p)[x] factors are removed, and the combinations loop is restarted from the beginning with the remaining Q[x] factors). It assumes that the order of elements in a defaultdict doesn't change due to deletions,  and further assumes that that order is maintained in each combination returned by itertools.combinations(). If itertools.combinations() returned sets instead of tuples, this wouldn't be an issue; we could do that ourselves if we thought it was necessary, but current versions of python provide the necessary guarantees.