Add *, /, ** operators to bitstrings
Multiplication of same-length bitstrings treats them as GF(2) polynomials mod x^l+1, where l is the length of each bitstring. The leftmost bit is the constant term. Note that this means that the polynomial 1 is represented by a bitstring whose int value has just the msb set.
Division of same-length bitstrings is the inverse of multiplication, and raises a ZeroDivisionError if the divisor, as a GF(2) polynomial, is not relatively prime to x^l+1.
Exponentiation is repeated multiplication for positive integer exponents, repeated division for negative exponents, and the aforementioned representation of 1 for exponent 0.
The only change to numfuns fixed a typo in a comment.