Correct counterintuitive multiplication
For a couple of days, we defined bitstring multiplication so that the leftmost bit, bit 0, was the constant bit in the GF2 polynomial interpretation of the bitstring. This would mean that a bitstring with int value 1 was NOT the polynomial 1. We're changing that to the much more intuitive interpretation of int value, with the lsb (bit -1) being the constant term. Note that the difference between these two interpretations is merely a one-position bit rotation of the product.